astar documentation and explanations
