+++ algos/common/generalized-astar/IGeneralizedAstarAutorouter.ts
@@ -1,0 +1,7 @@
+import type { Node, Point, PointWithObstacleHit } from "./types"
+
+export interface IGeneralizedAstarAutorouter {
+  computeG(current: Node, neighbor: Point): number
+  computeH(node: Point): number
+  getNeighbors(node: Node): Array<PointWithObstacleHit>
+}
